# marauders-map/contracts/openapi.yml
openapi: 3.1.0
info:
  title: The Marauder's Map API
  version: 1.0.0
  description: |
    Dark incident tracking system API - shared contract across all three implementations.

    ## The Three Houses
    - **🦁 Gryffindor**: React + Express (http://localhost:4001)
    - **🐍 Slytherin**: Angular + .NET (http://localhost:4002)
    - **🦅 Ravenclaw**: Spring Boot + Java (http://localhost:4003)

    All three implementations share the same PostgreSQL database and implement this exact API contract.

  contact:
    name: Hogwarts Ministry API Team
    email: api@hogwarts.edu

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4001
    description: Gryffindor (Express)
  - url: http://localhost:4002
    description: Slytherin (.NET)
  - url: http://localhost:4003
    description: Ravenclaw (Spring Boot)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login
  schemas:
...
    User:
      type: object
      required:
        - id
        - email
        - role
        - firstName
        - lastName
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "harry.potter@hogwarts.edu"
        role:
          $ref: '#/components/schemas/UserRole'
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          example: "Harry"
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          example: "Potter"
        house:
          $ref: '#/components/schemas/HouseName'
        avatarUrl:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
    Incident:
      type: object
      required:
        - id
        - title
        - severity
        - location
        - status
        - reportedBy
        - reportedAt
      properties:
        id:
          type: integer
          format: int64
          example: 42
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: "Dark Mark over Hogwarts"
        description:
          type: string
          nullable: true
          example: "Multiple witnesses report seeing the Dark Mark in the sky above the castle."
        severity:
          $ref: '#/components/schemas/SeverityLevel'
        location:
          $ref: '#/components/schemas/LocationType'
        status:
          $ref: '#/components/schemas/IncidentStatus'
        reportedBy:
          $ref: '#/components/schemas/User'
        reportedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        resolvedBy:
          allOf:
            - $ref: '#/components/schemas/User'
            - nullable: true
        tags:
          type: array
          items:
            type: string
          example: ["investigation", "urgent"]
        evidenceUrls:
          type: array
          items:
            type: string
            format: uri
        witnessCount:
          type: integer
          minimum: 0
          example: 3
...
    UserRole:
      type: string
      enum:
        - STUDENT
        - PREFECT
        - AUROR
      description: |
        - `STUDENT`: Can view MISCHIEF incidents only
        - `PREFECT`: Can view/manage up to DANGEROUS
        - `AUROR`: Full access including UNFORGIVABLE
    HouseName:
      type: string
      enum:
        - GRYFFINDOR
        - SLYTHERIN
        - RAVENCLAW
        - HUFFLEPUFF
      nullable: true
    SeverityLevel:
      type: string
      enum:
        - MISCHIEF
        - SUSPICIOUS
        - DANGEROUS
        - UNFORGIVABLE
      description: |
        - `MISCHIEF`: Minor rule breaking, pranks
        - `SUSPICIOUS`: Questionable behavior worth monitoring
        - `DANGEROUS`: Serious threat requiring immediate attention
        - `UNFORGIVABLE`: Dark magic, highest severity
    LocationType:
      type: string
      enum:
        - HOGWARTS
        - HOGSMEADE
        - KNOCKTURN_ALLEY
        - FORBIDDEN_FOREST
        - MINISTRY
        - AZKABAN
        - DIAGON_ALLEY
        - PLATFORM_9_3_4
    IncidentStatus:
      type: string
      enum:
        - OPEN
        - IN_PROGRESS
        - RESOLVED
        - ARCHIVED
...
    CreateIncidentRequest:
      type: object
      required:
        - title
        - severity
        - location
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        severity:
          $ref: '#/components/schemas/SeverityLevel'
        location:
          $ref: '#/components/schemas/LocationType'
        tags:
          type: array
          items:
            type: string
    UpdateIncidentRequest:
      type: object
      properties:
        severity:
          $ref: '#/components/schemas/SeverityLevel'
        status:
          $ref: '#/components/schemas/IncidentStatus'
        description:
          type: string
    PaginatedIncidents:
      type: object
      required:
        - data
        - total
        - page
        - limit
        - hasMore
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Incident'
        total:
          type: integer
          description: Total count of incidents matching filters
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        hasMore:
          type: boolean
...
paths:
...
  /api/incidents:
    get:
      tags: [Incidents]
      summary: List incidents with filtering and pagination
      operationId: listIncidents
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          description: Full-text search on title, description, location
          schema:
            type: string
        - name: severity
          in: query
          description: Filter by severity (comma-separated)
          schema:
            type: array
            items:
              $ref: '#/components/schemas/SeverityLevel'
          style: form
          explode: false
        - name: location
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/LocationType'
          style: form
          explode: false
        - name: status
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/IncidentStatus'
          style: form
          explode: false
        - name: sort
          in: query
          schema:
            type: string
            enum: [reportedAt, severity, status]
            default: reportedAt
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedIncidents'
        '401':
          description: Unauthenticated
        '403':
          description: Insufficient permissions

    post:
      tags: [Incidents]
      summary: Create new incident
      operationId: createIncident
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncidentRequest'
      responses:
        '201':
          description: Incident created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          description: Validation error
        '401':
          description: Unauthenticated
  /api/incidents/{id}:
    get:
      tags: [Incidents]
      summary: Get incident by ID
      operationId: getIncident
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '404':
          description: Incident not found
        '403':
          description: Insufficient permissions

    patch:
      tags: [Incidents]
      summary: Update incident
      operationId: updateIncident
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIncidentRequest'
      responses:
        '200':
          description: Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '403':
          description: Insufficient permissions
        '404':
          description: Incident not found

    delete:
      tags: [Incidents]
      summary: Delete incident (soft delete)
      operationId: deleteIncident
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted successfully
        '403':
          description: Insufficient permissions (AUROR only)
        '404':
          description: Incident not found

  # ===========================================================================
  # Comments
  # ===========================================================================
...